requirements.txt
annotated-types==0.7.0 
anyio==4.12.1 
cachetools==6.2.4 
certifi==2026.1.4 
cffi==2.0.0 
charset_normalizer==3.4.4 
click==8.3.1 
colorama==0.4.6 
cryptography==46.0.3 
deprecation==2.1.0 
fsspec==2026.1.0 
greenlet==3.3.0 
h11==0.16.0 
h2==4.3.0 
hpack==4.1.0 
httpcore==1.0.9 
httpx==0.28.1 
hyperframe==6.1.0 
idna==3.11 
markdown-it-py==4.0.0 
mdurl==0.1.2 
mmh3==5.2.0 
multidict==6.7.0 
packaging==25.0 
postgrest==2.27.2 
propcache==0.4.1 
psycopg2-binary==2.9.11 
pycparser==2.23 
pydantic==2.12.5 
pydantic-core==2.41.5 
pydantic-settings==2.12.0 
pygments==2.19.2 
pyiceberg==0.10.0 
pyjwt==2.10.1 
pyparsing==3.3.1 
pyroaring==1.0.3 
python-dateutil==2.9.0.post0 
python-dotenv==1.2.1 
realtime==2.27.2 
requests==2.32.5 
rich==14.2.0 
six==1.17.0 
sortedcontainers==2.4.0 
sqlalchemy==2.0.45 
storage3==2.27.2 
strenum==0.4.15 
strictyaml==1.7.3 
supabase==2.27.2 
supabase-auth==2.27.2 
supabase-functions==2.27.2 
tenacity==9.1.2 
typing-extensions==4.15.0 
typing-inspection==0.4.2 
urllib3==2.6.3 
websockets==15.0.1 
yarl==1.22.0
Mako==1.3.10 
MarkupSafe==3.0.3 
alembic==1.18.1
annotated-doc==0.0.4 
fastapi==0.128.0 
httptools==0.7.1 
pyyaml==6.0.3 
starlette==0.50.0 
uvicorn==0.40.0 
watchfiles==1.1.1
numpy==2.4.1 
pgvector==0.4.2
beautifulsoup4==4.13.5 
colorclass==2.2.2 
compressed==rtf-1.0.7 
distro==1.9.0 
easygui==0.98.3 
ebcdic==1.1.1 
extract-msg==0.55.0 
google-auth==2.47.0 
google-genai==1.60.0 
lark==1.3.1 
msoffcrypto-tool==6.0.0 
olefile==0.47 
oletools==0.60.2 
pcodedmp==1.2.6 
pyasn1==0.6.2 
pyasn1-modules==0.4.2 
red-black-tree-mod==1.22 
rsa==4.9.1 
rtfde==0.1.2.2 
sniffio==1.3.1 
soupsieve==2.8.3 
tzdata==2025.3 
tzlocal==5.3.1 
win-unicode-console==0.5